requirements.txt
absl-py==0.7.1
alabaster==0.7.10
astor==0.7.1
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
cloudpickle==1.0.0
cycler==0.10.0
decorator==4.4.0
docutils==0.14
gast==0.2.2
gin-config==0.1.3
google-pasta==0.1.6
grpcio==1.20.1
h5py==2.9.0
idna==2.6
imagesize==1.0.0
Jinja2==2.10
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
MarkupSafe==1.0
mock==3.0.5
numpy==1.16.3
packaging==17.1
pbr==5.1.3
protobuf==3.7.1
pygame==1.9.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
seaborn==0.8.1
six==1.12.0
snowballstemmer==1.2.1
sphinxcontrib-websupport==1.0.1
tb-nightly==1.14.0a20190514
tensorboard==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tf-agents-nightly==0.2.0.dev20190430
tf-estimator-nightly==1.14.0.dev2019051401
tf-nightly==1.14.1.dev20190514
tfp-nightly==0.7.0.dev20190514
urllib3==1.22
Werkzeug==0.15.2
wrapt==1.11.1
